refactor: Changed dir of components to root folder, they should have been there
