+++ src/core/components/IComponentData.ts
