+++ src/core/components/baseComponent.ts
@@ -1,5 +1,5 @@
-import type { Shader } from "../../gl/shaders";
-import type { SimObject } from "../../world/simObject";
+import type { Shader } from "../gl/shaders";
+import type { SimObject } from "../world/simObject";
 import type { IComponent } from "./IComponent";
 import type { IComponentData } from "./IComponentData";
