+++ src/core/components/componentManager.ts
