+++ src/core/components/spriteComponent.ts
@@ -1,6 +1,6 @@
-import type { Shader } from "../../gl/shaders";
-import { Sprite } from "../../graphics/sprite";
-import { Vector3 } from "../../math/vector3";
+import type { Shader } from "../gl/shaders";
+import { Sprite } from "../graphics/sprite";
+import { Vector3 } from "../math/vector3";
 import { BaseComponent } from "./baseComponent";
 import { ComponentManager } from "./componentManager";
 import type { IComponent } from "./IComponent";
